refactor(audio): move Apple audio enums to public audio types
Relocate AppleAudioCategory / AppleAudioCategoryOption / AppleAudioMode from
the internal wire file (support/native_audio.dart) into audio_session.dart,
where the public AppleAudioSessionConfiguration already lives. native_audio.dart
keeps only the internal NativeAudioConfiguration wire type and its toStringValue
serialization, importing the enums from the audio layer. Removes the
import-plus-export of native_audio from audio_session.dart. Public API and
behavior are unchanged.
